# Generate `flink-conf.yaml` file automatically to set optimum conf values