--- a/docker/.env
+++ b/docker/.env
@@ -21,3 +21,6 @@ DWH_ROOT=./dwh
 # This is for openmrs-compose.yaml; for big data dump set `big`, for small
 # data dump set `small`, which is the default mode.
 DATABASE_DUMP_MODE=small
+
+# JVM parameters to be used by the application
+JAVA_OPTS=-Xms2g -Xmx2g