--- a/pipelines/controller/src/main/java/com/google/fhir/analytics/DataProperties.java
+++ b/pipelines/controller/src/main/java/com/google/fhir/analytics/DataProperties.java
@@ -93,6 +93,8 @@ public class DataProperties {
 
   private String fhirServerUserName;
 
+  private boolean autoGenerateFlinkConfiguration;
+
   private String fhirServerOAuthTokenEndpoint;
 
   private String fhirServerOAuthClientId;