## Description of what I changed

Fixes #823 
- Enabled a feature flag to automate the generation of `flink-conf.yaml` based on the `numThreads` in `application.yaml` file and the number of `cores` available in the machine.
- If the flag is disabled, then the `flink-conf.yaml` directed by the env. var `FLINK_CONF_DIR` will be used. If var is empty, then default conf will be used.
- JAVA_OPTS can be set to override the default JVM heap and other parameter values.
-  The application will fail to launch if the JVM Off Heap memory is insufficient for the Flink cluster to launch.
- Fixes the regression introduced by this [change](https://github.com/google/fhir-data-pipes/pull/670) where the maxWorkers was defaulted to 1 and in turn the num of shards for parquet files was also 1.

## E2E test

Tested `Full Run` and `Incremental Run` by enabling and disabling the flags. Also tested for different `numThread` configurations.

TESTED:

Load tested the application on a 48 core machine. Given below are the results

Run # | Run Type | # of Patient Records | Time in secs (Before Change) | Time in secs (After Change) | Remarks
-- | -- | -- | -- | -- | --
1 | Full | 9K | 400 | 358 | Full run timing remains the same before and after the changes
2 | Full (Repeat) | 9K | 360 | 256 |  
3 | Incremental | 9K Upfront + 100 additional | 140 | 30 | The improvement in time after change is because of fixing the regression (increased num of shards from 1 to N)
4 | Incremental | 9K Upfront + 500 additional | 180 | 32 |  

After fixing the regression the time taken for incremental run has significantly reduced

## Checklist: I completed these to help reviewers :)

<!--- Put an `x` in the box if you did the task -->
<!--- If you forgot a task please follow the instructions below -->

- [x] I have read and will follow the [review process](https://github.com/GoogleCloudPlatform/openmrs-fhir-analytics/blob/master/doc/review_process.md).
- [x] I am familiar with Google Style Guides for the language I have coded in.

  No? Please take some time and review [Java](https://google.github.io/styleguide/javaguide.html) and [Python](https://google.github.io/styleguide/pyguide.html) style guides.

- [x] My IDE is configured to follow the Google [**code styles**](https://google.github.io/styleguide/).

  No? Unsure? -> [configure your IDE](https://github.com/google/google-java-format).

- [x] I have **added tests** to cover my changes. (If you refactored existing code that was well tested you do not have to add tests)
- [x] I ran `mvn clean package` right before creating this pull request and added all formatting changes to my commit.
- [x] All new and existing **tests passed**.
- [x] My pull request is **based on the latest changes** of the master branch.

  No? Unsure? -> execute command `git pull --rebase upstream master`