@chandrashekar-s 2023-11-28
Thanks @bashir2 for the reviewing the changes. I have addressed some of the comments in the latest commit. Also I have created these 2 issues #893 and #891 for `Validating Flink in non-local mode` and `Inestigating/fixing the reshuffle operations for writing to parquet files` respectively.