@chandrashekar-s 2023-11-29
The performance results have been attached in the PR description. No noticeable changes for the `Full run`, but for `Incremental run` the timing has improved after fixing the number of Shards to be `N` (Flink parallelism)